requirements.txt
streamlit
torch
torchvision
clip
matplotlib
Pillow
tqdm
numpy
faiss-cpu
sentence-transformers
git+https://github.com/openai/CLIP.git